+++ RELEASE-NOTES.md
@@ -1,3 +1,8 @@
+### 2.5.2
+* Fixing log level that stopped of work after **2.5.0**
+* Fixing and increasing docs development instructions
+* Fixing wildcard resolution were not solving main domain to docker container, just the subdomains
+
 ### 2.5.1
 * Fixing ping slowness, takes more than 10 seconds to respond